requirements.txt
torch>=2.0
torchvision
timm
ultralytics
wandb
Pillow
numpy
tqdm
huggingface_hub
datasets
